# core/docs/swagger.yaml
definitions:
...
  apprun_modules_auth_service.LoginRequest:
    properties:
      identifier:
        description: Username or email
        example: user@example.com
        type: string
      password:
        description: Password
        example: SecurePass123
        type: string
    required:
    - identifier
    - password
    type: object
  apprun_modules_auth_service.LoginResponse:
    properties:
      access_token:
        description: JWT access token (short-lived)
        type: string
      expires_in:
        description: Seconds until access token expires
        type: integer
      refresh_token:
        description: JWT refresh token (long-lived)
        type: string
      user:
        allOf:
        - $ref: '#/definitions/apprun_modules_auth_service.UserProfile'
        description: User profile data
    type: object
...
  apprun_modules_auth_service.UserProfile:
    properties:
      created_at:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: string
      language:
        type: string
      nickname:
        type: string
      phone:
        type: string
      status:
        type: integer
      timezone:
        type: string
      username:
        type: string
    type: object
  apprun_pkg_response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  apprun_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        $ref: '#/definitions/apprun_pkg_response.ErrorInfo'
      message:
        type: string
      request_id:
        type: string
      success:
        type: boolean
    type: object
...
  modules_auth_handler.CheckPermissionRequest:
    properties:
      action:
        type: string
      resource:
        type: string
    required:
    - action
    - resource
    type: object
  modules_auth_handler.CheckPermissionResponse:
    properties:
      allowed:
        type: boolean
    type: object
...
  modules_auth_handler.PermissionItem:
    properties:
      actions:
        items:
          type: string
        type: array
      resource:
        type: string
    type: object
  modules_auth_handler.PermissionsResponse:
    properties:
      permissions:
        items:
          $ref: '#/definitions/modules_auth_handler.PermissionItem'
        type: array
    type: object
...
info:
  contact:
    email: support@websoft9.com
    name: API Support
  description: AppRun Platform REST API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: AppRun API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apprun_modules_auth_service.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/apprun_modules_auth_service.LoginResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "403":
          description: Account disabled
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      summary: User login
      tags:
      - auth
  /api/auth/me:
    get:
      consumes:
      - application/json
      description: Retrieve profile of the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/apprun_modules_auth_service.UserProfile'
              type: object
        "401":
          description: Missing or invalid token
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - auth
...
  /api/projects/{project_id}/permissions/check:
    post:
      consumes:
      - application/json
      description: Check if the current user has a specific permission in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Permission to check
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/modules_auth_handler.CheckPermissionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/modules_auth_handler.CheckPermissionResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Check permission
      tags:
      - rbac
  /api/projects/{project_id}/permissions/me:
    get:
      description: Get the current user's permissions in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/apprun_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/modules_auth_handler.PermissionsResponse'
              type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/apprun_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get my permissions
      tags:
      - rbac
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"